retreived categories (#4)
* updated endpointed

changed endpoint to the correct

* updated endpointed

changed endpoint to the correct

* Queried categories

WPrequest.js, queries the WordPress API to retrieve all categories from The Hoya and exports the data to a CSV file. The CSV contains category names, IDs, and associated href links, making it easy to fetch articles by category. we can use the links to request articles for specific Categories such as Featured News, Opinion, and Guide a. The next step in enhancing the script is to build functions that take a category ID as a parameter and return an array of objects containing article content, headers, and related metadata for each retrieved article. This will streamline the process of gathering and organizing articles based on categories.

 MUST INSTALL :npm install json2csv in server folder b4 running